Time: 7 ms (18.03%) | Memory: 12.6 MB (89.28%) - LeetSync

+++ 2102-find-the-middle-index-in-array/find-the-middle-index-in-array.cpp
@@ -4,9 +4,9 @@ public:
         int n=nums.size();
         if(n==1)
         return 0;
-        long sum=accumulate(nums.begin(),nums.end(),0);
+        int sum=accumulate(nums.begin(),nums.end(),0);
        
-        long tmp=0;
+        int tmp=0;
         for(int i=0;i<n;i++){
             if(i!=0)
             tmp=tmp+nums[i-1];
